refactor(tui): split app.go into per-domain message handlers
Drop internal/tui/app.go from 5,232 to ~1,400 lines by relocating each
case block in Update() and each domain-specific helper into a focused
file:

- update_lifecycle.go  initApp / tick / agentEvent / createResult / killResult / diffStats / resumeDone / windowSize
- update_plan.go       plannerQuestion / planEditor*Msg / openPlanEditor / approvePlanAndSpawn / submitPromptModal
- update_pr.go         prDraftReady / prCreated / prPoll / mergePR / pollAllSessions / startPRDraftCmd / submitPRComposeModal / openURL
- update_review.go     reviewDiff / reviewVerdict / fetchReviewDiffCmd / reviewTaskCmd / handleReviewReworkRequest / handleShippingFeedbackRequest / buildFeedbackPrompt / buildReviewReworkPrompt
- update_keys.go       updateDashboard / updateFocusLaunchKeys / closeFocusLaunchAgent
- update_pickers.go    updateFileBrowser / updateBranchPicker / updateRepoPicker
- update_views.go      updateDiff / updateGlobalConfig / returnFromConfigForm / initRepoConfigForm / extractRepoSettings

Update() is now a thin type-switch dispatcher; each handleXMsg returns
(tea.Model, tea.Cmd). handleWindowSize returns App so the existing
intentional fall-through to the view router is preserved.

Pure mechanical refactor. No user-visible behavior change.
go test -race ./internal/tui/... is green.
