+++ internal/tui/update_views.go
@@ -1,0 +1,171 @@
+package tui
+
+import (
+	tea "charm.land/bubbletea/v2"
+	"github.com/devenjarvis/refrain/internal/config"
+)
+
+func (a App) returnFromConfigForm() (tea.Model, tea.Cmd) {
+	if a.repoPickerPendingFromConfig {
+		a.repoPickerPendingFromConfig = false
+		counts := make(map[string]int, len(a.cfg.Repos))
+		for _, repo := range a.cfg.Repos {
+			if mgr := a.managers[repo.Path]; mgr != nil {
+				counts[repo.Path] = mgr.AgentCount()
+			}
+		}
+		a.repoPicker.setRepos(a.cfg.Repos, counts, a.dashboard.configRepoPath)
+		a.dashboard.repoConfigForm = nil
+		a.dashboard.configRepoPath = ""
+		a.view = ViewRepoPicker
+		return a, nil
+	}
+	a.dashboard.panelFocus = focusList
+	a.dashboard.repoConfigForm = nil
+	return a, nil
+}
+func (a App) updateDiff(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg.(type) {
+	case diffCloseMsg:
+		a.view = ViewDashboard
+		return a, nil
+	}
+
+	var cmd tea.Cmd
+	a.diff, cmd = a.diff.Update(msg)
+	return a, cmd
+}
+func (a *App) initRepoConfigForm(repoPath string) {
+	rs := a.repoSettings[repoPath]
+	if rs == nil {
+		rs = &config.RepoSettings{}
+	}
+
+	bypassPerms := config.DefaultBypassPermissions
+	if rs.BypassPermissions != nil {
+		bypassPerms = *rs.BypassPermissions
+	}
+	defaultBranch := ""
+	if rs.DefaultBranch != nil {
+		defaultBranch = *rs.DefaultBranch
+	}
+	branchPrefix := ""
+	if rs.BranchPrefix != nil {
+		branchPrefix = *rs.BranchPrefix
+	}
+	agentProgram := ""
+	if rs.AgentProgram != nil {
+		agentProgram = *rs.AgentProgram
+	}
+	planModel := ""
+	if rs.PlanModel != nil {
+		planModel = *rs.PlanModel
+	}
+	agentModel := ""
+	if rs.AgentModel != nil {
+		agentModel = *rs.AgentModel
+	}
+	ideCommand := ""
+	if rs.IDECommand != nil {
+		ideCommand = *rs.IDECommand
+	}
+	worktreeDir := ""
+	if rs.WorktreeDir != nil {
+		worktreeDir = *rs.WorktreeDir
+	}
+	alias := ""
+	for _, r := range a.cfg.Repos {
+		if r.Path == repoPath {
+			alias = r.Alias
+			break
+		}
+	}
+
+	inputWidth := 30
+	var fields []formField
+	fields = addTextInput(fields, "Alias", alias, "short nickname", inputWidth)
+	fields = addToggle(fields, "Bypass Permissions", bypassPerms)
+	fields = addTextInput(fields, "Default Branch", defaultBranch, "auto-detect", inputWidth)
+	fields = addTextInput(fields, "Branch Prefix", branchPrefix, config.DefaultBranchPrefix, inputWidth)
+	fields = addTextInput(fields, "Agent Program", agentProgram, config.DefaultAgentProgram, inputWidth)
+	fields = addSelect(fields, "Plan Model", config.KnownModels, optionIndex(config.KnownModels, planModel))
+	fields = addSelect(fields, "Agent Model", config.KnownAgentModels, optionIndex(config.KnownAgentModels, agentModel))
+	fields = addEditorFields(fields, ideCommand)
+	fields = addTextInput(fields, "Worktree Directory", worktreeDir, config.DefaultWorktreeDir, inputWidth)
+
+	form := newConfigForm(fields, a.dashboard.fixedTermWidth())
+	a.dashboard.repoConfigForm = &form
+	a.dashboard.configRepoPath = repoPath
+	a.dashboard.panelFocus = focusConfig
+}
+
+// extractRepoSettings reads form values and creates a RepoSettings struct.
+func (a App) extractRepoSettings() *config.RepoSettings {
+	form := a.dashboard.repoConfigForm
+	if form == nil {
+		return &config.RepoSettings{}
+	}
+	s := &config.RepoSettings{}
+
+	bypassPerms := form.toggleValue("Bypass Permissions")
+	s.BypassPermissions = &bypassPerms
+
+	if v := form.textValue("Default Branch"); v != "" {
+		s.DefaultBranch = &v
+	}
+	if v := form.textValue("Branch Prefix"); v != "" {
+		s.BranchPrefix = &v
+	}
+	if v := form.textValue("Agent Program"); v != "" {
+		s.AgentProgram = &v
+	}
+	if v := form.selectValue("Plan Model"); v != "" {
+		s.PlanModel = &v
+	}
+	if v := form.selectValue("Agent Model"); v != "" {
+		s.AgentModel = &v
+	}
+	if v := extractIDECommand(*form); v != "" {
+		s.IDECommand = &v
+	}
+	if v := form.textValue("Worktree Directory"); v != "" {
+		s.WorktreeDir = &v
+	}
+	return s
+}
+func (a App) updateGlobalConfig(msg tea.Msg) (tea.Model, tea.Cmd) {
+	switch msg := msg.(type) {
+	case globalConfigSaveMsg:
+		// Persist global settings.
+		if err := config.SaveGlobalSettings(msg.settings); err != nil {
+			a.setError(err.Error())
+			a.view = ViewDashboard
+			return a, nil
+		}
+		a.globalSettings = msg.settings
+		// Rebuild resolved cache and push to all managers.
+		for repoPath, rs := range a.repoSettings {
+			a.resolvedCache[repoPath] = config.Resolve(a.globalSettings, rs)
+			if mgr := a.managers[repoPath]; mgr != nil {
+				mgr.UpdateSettings(a.resolvedCache[repoPath])
+			}
+		}
+		newResolved := config.Resolve(a.globalSettings, nil)
+		if newResolved.SidebarWidth != a.dashboard.sidebarWidth {
+			a.dashboard.sidebarWidth = newResolved.SidebarWidth
+			a.resizeAllForDashboard()
+		}
+		// Refresh wellness settings from updated global config.
+		a.focusSessionMinutes = newResolved.FocusSessionMinutes
+		a.focusBreakMinutes = newResolved.FocusBreakMinutes
+		a.view = ViewDashboard
+		return a, nil
+	case globalConfigCancelMsg:
+		a.view = ViewDashboard
+		return a, nil
+	}
+
+	var cmd tea.Cmd
+	a.globalConfig, cmd = a.globalConfig.Update(msg)
+	return a, cmd
+}
